feat: rename Linode to Akamai Connected Cloud
